+++ packages/lu/src/commands/qnamaker/convert.ts
@@ -7,9 +7,12 @@ import {CLIError, Command, flags, utils} from '@microsoft/bf-cli-command'
 const exception = require('./../../parser/lufile/classes/exception')
 const fs = require('fs-extra')
 const file = require('./../../utils/filehelper')
-const luConverter = require('./../../parser/converters/qnatoqnajsonconverter')
-const qnaConverter = require('./../../parser/converters/qnajsontoqnaconverter')
 const fileExtEnum = require('./../../parser/lufile/helpers').FileExtTypeEnum
+
+const QnAMaker = require('./../../parser/qna/qnamaker/qnamaker')
+const Alterations = require('./../../parser/qna/alterations/alterations')
+const QnAMakerBuilder = require('./../../parser/qna/qnamaker/qnaMakerBuilder')
+const alterationsBuilder = require('./../../parser/qna/alterations/alterationsBuilder')
 
 export default class QnamakerConvert extends Command {
   static description = 'Converts .lu file(s) to QnA application JSON models or vice versa.'
@@ -41,10 +44,16 @@ export default class QnamakerConvert extends Command {
       let result: any
       if (isQnA) {
         let luFiles = await file.getLuObjects(stdin, flags.in, flags.recurse, fileExtEnum.QnAFile)
-        result = await luConverter.parseQnaToJson(luFiles, false, flags.luis_culture)
+        result = {}
+        result.finalQnAJSON = await QnAMakerBuilder.build(luFiles, false, flags.luis_culture)
+        result.finalQnAAlterations = await alterationsBuilder.build(luFiles, false, flags.luis_culture)
       } else {
-        const qnAFile = stdin ? stdin : await file.getContentFromFile(flags.in)
-        result = await qnaConverter.parseQnAObjectToLu(qnAFile, flags.sort, flags.alterations)
+        const qnaContent = stdin ? stdin : await file.getContentFromFile(flags.in)
+        let QnA = flags.alterations ? new Alterations(file.parseJSON(qnaContent, 'QnA Alterations')) : new QnAMaker(file.parseJSON(qnaContent, 'QnA'))
+        if (flags.sort) {
+          QnA.sort()
+        }
+        result = QnA.parseToLuContent()
       }
 
       // If result is null or undefined return
@@ -69,6 +78,7 @@ export default class QnamakerConvert extends Command {
         }
       }
     } catch (err) {
+      // console.log(err)
       if (err instanceof exception) {
         throw new CLIError(err.text)
       }
