+++ packages/lu/test/parser/lufile/parseFileContents.modelAsFeature.test.js
@@ -3,7 +3,6 @@
  * Licensed under the MIT License.
  */
 const parseFile = require('../../../src/parser/lufile/parseFileContents');
-const validateLUISBlob = require('../../../src/parser/luisfile/parseLuisFile').validateLUISBlob;
 var chai = require('chai');
 var assert = chai.assert;
 describe('Model as feature definitions', function () {
